backend/requirements.txt
fastapi
uvicorn
pydantic
openai
azure-cognitiveservices-speech
python-multipart
python-dotenv
requests